Delete label from vector metric when deregistered
